KEYBOARD MOVEMENT CRISP AS IT COMES
Oooooooo wheeee this keyboard movement looks hot. I commented out the
transtition settings on style in client/src/blob.js also we have merged
the keyDown / keyUp inputs and handle them in the same function in
mainPlayer. handling them with the same function was the answer to a lot
of my weird movement problems and now everything looks sick. Pulling to
master.

This handles enhancement #5