Socket Frame Message Optimization
Reduced the size of messages being sent between client and server for
increased scalability. If you want to know all the abbrevations see the
abv_schema.txt. Also I added a round function to /models/vector and
/client/src/blobs/blob to use when sending messages that brings msg size
down a lot. we went from 70-80byte messages to 40-50. Also changed the
updater so that it only sends updates when something actually changes.
We'll have to keep building on that.